+++ demo/javascript/src/entry.js
@@ -59,10 +59,13 @@ Demo.conn = new WebIM.connection({
     isMultiLoginSessions: WebIM.config.isMultiLoginSessions,
     https: typeof WebIM.config.https === 'boolean' ? WebIM.config.https : location.protocol === 'https:',
     url: WebIM.config.xmppURL,
-    isAutoLogin: false,
     heartBeatWait: WebIM.config.heartBeatWait,
     autoReconnectNumMax: WebIM.config.autoReconnectNumMax,
-    autoReconnectInterval: WebIM.config.autoReconnectInterval
+    autoReconnectInterval: WebIM.config.autoReconnectInterval,
+    apiUrl: WebIM.config.apiURL,
+    isHttpDNS: WebIM.config.isHttpDNS,
+    isWindowSDK: WebIM.config.isWindowSDK,
+    isAutoLogin: false
 });
 
 Demo.api.render(document.getElementById('demo'));